almost done

+++ ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: noah-baz <noah-baz@student.42.fr>          +#+  +:+       +#+        */
+/*   By: nbaz-sil <nbaz-sil@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/22 16:47:47 by nbaz-sil          #+#    #+#             */
-/*   Updated: 2026/06/06 06:49:33 by noah-baz         ###   ########.fr       */
+/*   Updated: 2026/06/10 20:30:58 by nbaz-sil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
 	int	ft_putstr(char *str); 
 	int	ft_putnbr(int n); 
 	int	ft_putptr(void *ptr);
-	int	ft_puthex(unsigned int num, char format);
+	int	ft_puthex(unsigned long num, char format);
 	int	ft_putuns(unsigned int num);
 
 #endif
