+++ ft_putptr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putptr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: noah-baz <noah-baz@student.42.fr>          +#+  +:+       +#+        */
+/*   By: nbaz-sil <nbaz-sil@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/22 17:01:11 by nbaz-sil          #+#    #+#             */
-/*   Updated: 2026/06/07 16:02:26 by noah-baz         ###   ########.fr       */
+/*   Updated: 2026/06/10 20:32:12 by nbaz-sil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,6 +21,6 @@ int	ft_putptr(void *str)
 		write(1, "(nil)", 5);
 	ptr = (unsigned long  *)str;
 	write(1, "0x", 2);
-    i = ft_puthex(ptr, 'x');
+    i = ft_puthex((unsigned long)ptr, 'x');
 	return (i + 2);
 }
